adapt to #[deny(transmute_from_fn_item_types)]

+++ src/internal/js/class.rs
@@ -252,7 +252,7 @@ pub trait ClassInternal: Class {
 
             let (call_callback, call_kernel) = match descriptor.call {
                 Some(k) => k.export(),
-                None    => (mem::transmute(ConstructorCallKernel::unimplemented::<Self>), null_mut())
+                None    => (mem::transmute(ConstructorCallKernel::unimplemented::<Self> as usize), null_mut())
             };
 
             let metadata_pointer = neon_sys::class::create_base(isolate,
